Change COFF signature from 32 to 16 bit.

+++ DLLDependencyViewer/DLLDependencyViewer/src/nogui/pe/coff_optional_standard.h
@@ -52,8 +52,8 @@ struct coff_optional_header_standard_32_64
 };
 
 
-static constexpr std::uint32_t const s_coff_optional_sig_32 = 0x010b;
-static constexpr std::uint32_t const s_coff_optional_sig_64 = 0x020b;
+static constexpr std::uint16_t const s_coff_optional_sig_32 = 0x010b;
+static constexpr std::uint16_t const s_coff_optional_sig_64 = 0x020b;
 
 
 e_pe_parse_coff_optional_header_standard_32_64 pe_parse_coff_optional_header_standard_32_64(void const* const& file_data, int const& file_size, coff_optional_header_standard_32_64 const*& header);
